refactor: Move project service into project dir
