+++ frontend/src/app/projects/models/backup-settings/trigger-pipeline/trigger-pipeline.component.ts
@@ -15,7 +15,7 @@ import {
   Model,
   ModelService,
 } from 'src/app/projects/models/service/model.service';
-import { ProjectService } from 'src/app/services/project/project.service';
+import { ProjectService } from 'src/app/projects/service/project.service';
 import { SessionService } from 'src/app/sessions/service/session.service';
 import { CreateBackupComponent } from '../create-backup/create-backup.component';
 import { BackupService, Pipeline } from '../service/backup.service';
