+++ frontend/src/app/projects/models/choose-init/choose-init.component.ts
@@ -5,7 +5,7 @@
 
 import { Component, EventEmitter, Output } from '@angular/core';
 import { ModelService } from 'src/app/projects/models/service/model.service';
-import { ProjectService } from 'src/app/services/project/project.service';
+import { ProjectService } from '../../service/project.service';
 
 @Component({
   selector: 'app-choose-init',
