+++ frontend/src/app/projects/models/model-description/model-description.component.ts
@@ -9,12 +9,12 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { combineLatest, filter, switchMap, tap } from 'rxjs';
 import { ToastService } from 'src/app/helpers/toast/toast.service';
 import { ModelService } from 'src/app/projects/models/service/model.service';
-import { ProjectService } from 'src/app/services/project/project.service';
 import {
   ToolNature,
   ToolService,
   ToolVersion,
 } from 'src/app/settings/core/tools-settings/tool.service';
+import { ProjectService } from '../../service/project.service';
 
 @Component({
   selector: 'app-model-description',
