+++ frontend/src/app/projects/models/model-restrictions/model-restrictions.component.ts
@@ -15,7 +15,7 @@ import {
   Model,
   ModelService,
 } from 'src/app/projects/models/service/model.service';
-import { ProjectService } from 'src/app/services/project/project.service';
+import { ProjectService } from '../../service/project.service';
 
 @Component({
   selector: 'app-model-restrictions',
