+++ frontend/src/app/projects/models/model-source/choose-source.component.ts
@@ -5,8 +5,8 @@
 
 import { Component, EventEmitter, Output } from '@angular/core';
 import { ModelService } from 'src/app/projects/models/service/model.service';
-import { ProjectService } from 'src/app/services/project/project.service';
 import { UserService } from 'src/app/services/user/user.service';
+import { ProjectService } from '../../service/project.service';
 
 @Component({
   selector: 'app-choose-source',
