+++ frontend/src/app/projects/project-detail/model-overview/model-overview.component.ts
@@ -10,14 +10,11 @@ import {
   ModelService,
 } from 'src/app/projects/models/service/model.service';
 import { ProjectUserService } from 'src/app/projects/project-detail/project-users/service/project-user.service';
-import {
-  Project,
-  ProjectService,
-} from 'src/app/services/project/project.service';
 import { UserService } from 'src/app/services/user/user.service';
 import { NewReadonlySessionDialogComponent } from 'src/app/sessions/new-readonly-session-dialog/new-readonly-session-dialog.component';
 import { SessionService } from 'src/app/sessions/service/session.service';
 import { TriggerPipelineComponent } from '../../models/backup-settings/trigger-pipeline/trigger-pipeline.component';
+import { Project, ProjectService } from '../../service/project.service';
 
 @Component({
   selector: 'app-model-overview',
