+++ frontend/src/app/projects/project-detail/project-metadata/project-metadata.component.ts
@@ -27,7 +27,7 @@ import {
   PatchProject,
   Project,
   ProjectService,
-} from 'src/app/services/project/project.service';
+} from '../../service/project.service';
 
 @Component({
   selector: 'app-project-metadata',
