+++ frontend/src/app/projects/project-wrapper/project-wrapper.component.ts
@@ -12,10 +12,7 @@ import {
   ModelService,
 } from 'src/app/projects/models/service/model.service';
 import { ProjectUserService } from 'src/app/projects/project-detail/project-users/service/project-user.service';
-import {
-  Project,
-  ProjectService,
-} from 'src/app/services/project/project.service';
+import { Project, ProjectService } from '../service/project.service';
 
 @Component({
   selector: 'app-project-wrapper',
