+++ frontend/src/app/settings/settings.component.ts
@@ -4,7 +4,7 @@
  */
 
 import { Component, OnInit } from '@angular/core';
-import { ProjectService } from 'src/app/services/project/project.service';
+import { ProjectService } from '../projects/service/project.service';
 import { UserService } from '../services/user/user.service';
 
 @Component({
